+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - Using Tool websearch doesn´t work anymore (https://github.com/nbonamy/witsy/issues/428)
 - For Ollama 'Max Completion Tokens' should allow negative values (https://github.com/nbonamy/witsy/issues/413)
 - Added keep-alive on ollama embeddings (https://github.com/nbonamy/witsy/issues/425)
+- Scopes from OAuth protected resource metadata should be used for DCR (https://github.com/nbonamy/witsy/issues/420)
 
 ### Removed
 - N/A